requirements.txt
aiogram
httpx
python-dotenv

